Refactored bot structures. Moved go code into bot/ directory.
